requirements.txt
torch
torchvision
torchaudio
matplotlib
numpy
pillow
tqdm
scipy
opencv-python-headless
streamlit
pyngrok
twilio
yagmail
